Refactor Repeat source and harden edge cases
Extracted Repeat implementation from nested mixin classes into reusable Advanced types (`RepeatSourceSignal`, `RepeatSourceCoordinator`, and `RepeatSourceWitness`) and promoted them into the shipped public API across target frameworks. The repeat coordinator now explicitly handles source subscribe exceptions as a single downstream error and avoids duplicate terminal/repeat scheduling behavior. Added Repeat parity tests for subscribe-throw, duplicate completion, and stale late terminal/error paths, and standardized repository line-ending policy to LF via `.editorconfig` and `.gitattributes`.
