+++ .editorconfig
@@ -7,6 +7,7 @@ root = true
 # Default settings
 #############################################
 [*]
+end_of_line = lf
 insert_final_newline = true
 indent_style = space
 indent_size = 4
